pyproject.toml
[tool.poetry]
name = "rest-chat"
version = "0.1.0"
description = ""
authors = ["Kyle Charlton <kyle@kcharlton.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.108.0"
uvicorn = "0.25.0"
pytest = "7.4.0"
httpx = "0.26.0"
sqlmodel = "0.0.16"
passlib = "1.7.4"
python-jose = "3.3.0"
bcrypt = "4.1.2"
cryptography = "42.0.2"
python-multipart = "^0.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
